Always use 64-bit integer for file size.
`size_t` depends on platform, meaning on 32-bit systems we can't store
file sizes over 2GiB if we use it.

+++ vfs.h
@@ -31,7 +31,7 @@ struct file_set {
 struct directory_entry {
 	enum file_type type;
 	char *name;
-	ssize_t size;
+	int64_t size;
 	uint64_t serial_no;
 };
